+++ l3experimental/CHANGELOG.md
@@ -7,6 +7,9 @@ this project uses date-based 'snapshot' version identifiers.
 
 ## [Unreleased]
 
+### Added
+- New `l3bitset` module
+
 ## [2020-10-27]
 
 ### Changed